backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ehon-backend"
version = "0.1.0"
description = "絵本生成アプリ Backend"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.23",
  "pydantic>=2.5",
  "openai>=1.3.0",
  "python-multipart>=0.0.6",
  "python-dotenv>=1.0.0",
  "supabase>=2.0.0",
  "httpx>=0.25.0",
]

[tool.uvicorn]
factory = false
